pi/pipeline_monitor/requirements.txt
apache-iotdb==1.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.0.0
idna==3.7
numpy==1.26.4
packaging==24.0
pandas==1.5.3
pkg_resources==0.0.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
RPi.GPIO==0.7.1
six==1.16.0
SQLAlchemy==1.4.52
SQLAlchemy-Utils==0.41.2
testcontainers==4.3.3
thrift==0.20.0
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
